Move replay recorder to final location
