+++ osu.Game.Rulesets.Osu/UI/OsuReplayRecorder.cs
@@ -1,0 +1,23 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Collections.Generic;
+using osu.Game.Replays;
+using osu.Game.Rulesets.Osu.Replays;
+using osu.Game.Rulesets.Replays;
+using osu.Game.Rulesets.UI;
+using osuTK;
+
+namespace osu.Game.Rulesets.Osu.UI
+{
+    public class OsuReplayRecorder : ReplayRecorder<OsuAction>
+    {
+        public OsuReplayRecorder(Replay replay)
+            : base(replay)
+        {
+        }
+
+        protected override ReplayFrame HandleFrame(Vector2 position, List<OsuAction> actions, ReplayFrame previousFrame)
+            => new OsuReplayFrame(Time.Current, position, actions.ToArray());
+    }
+}
